Move failure related tests into separate module
